% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biol280data-package.R
\docType{package}
\name{biol280data-package}
\alias{biol280data}
\alias{biol280data-package}
\title{biol280data: Automatically Makes All Gapminder Data Sets (And A Few Others) Available in R.}
\description{
This package makes all of the datasets from Gapminder available automatically in R. It replaces the manual process of downloadin Gapminder data separately and then uploading it to R. As additional non-Gapminder data sets become available, they will be added.
}
\author{
\strong{Maintainer}: Jeff Wesner \email{jeff.wesner@usd.edu} (\href{https://orcid.org/0000-0001-6058-7972}{ORCID})

}
\keyword{internal}
